Database integration for the available versions part of OTA
